Game dev vite

+++ traveler/src/stores/UserStore.ts
@@ -1,27 +1,29 @@
-import {defineStore} from "pinia";
-import UserModel from "@/models/UserModel";
-import * as Request from '@/api/Requests'
+import {defineStore} from 'pinia';
+import UserModel from '@/models/UserModel';
+import * as Request from '@/api/Requests';
 import router from '@/router/index';
-import {useHeroStore} from '@/stores/HeroStore'
-import {useBagStore} from '@/stores/BagStore'
-import {useMapStore} from '@/stores/MapStore'
+import {useHeroStore} from '@/stores/HeroStore';
+import {useMapLocationStore} from '@/stores/map-location-store';
+import {useBagStore} from '@/stores/BagStore';
+import {useChestStore} from '@/stores/ChestStore';
+import MapModel from "@/models/MapModel";
 
-export const useUserStore = defineStore("user", {
-
-    state: () => {
-        return {
-            user: new UserModel().build(),
-            error: "",
-        };
-    },
+export const useUserStore = defineStore('user', {
+    state: () => ({
+        user: new UserModel().build(),
+        error: '',
+    }),
     getters: {
-        isLoggedIn: (() => localStorage.getItem("uStatus") === "true")
+        isLoggedIn: (): boolean => localStorage.getItem('uStatus') === 'true',
+        isUserLoggedIn(): boolean {
+            return this.user.getStatus();
+        }
     },
     actions: {
         async login(username: string, password: string) {
             const userFromApi = await Request.login(username, password);
             if (userFromApi == null) {
-                this.error = username + " is not found."
+                this.error = `${username} is not found.`;
                 return false;
             }
             this.user
@@ -30,35 +32,49 @@ export const useUserStore = defineStore("user", {
                 .setId(userFromApi.id)
                 .setLoggedIn(true);
 
-            localStorage.setItem("uStatus", this.user.getStatus().toString());
+            localStorage.clear();
+            localStorage.setItem('uStatus', 'true');
 
             const heroStore = useHeroStore();
             heroStore.getHero();
 
-            const mapStore = useMapStore();
-            await mapStore.resetMap(); // Reset map on login
-
-            const bagStore = useBagStore();
-            await bagStore.resetBag();
-            this.error = "";
-            router.push("/home")
-
+            this.error = '';
+            router.push('/home');
             return true;
         },
         async logout(): Promise<void> {
             console.log('Logout initiated');
-            const mapStore = useMapStore();
+
+            const mapLocationStore = useMapLocationStore();
             const bagStore = useBagStore();
-            await mapStore.resetMap().then(async () => {
-                await bagStore.resetBag().then(() => {
-                    localStorage.removeItem("map");
+            const chestStore = useChestStore();
+
+            mapLocationStore.initMapsList();
+
+            const oldForestMap: MapModel | undefined = mapLocationStore.mapsList.find(map => map.name === 'Old Forest') as MapModel;
+
+            if (oldForestMap) {
+                try {
+                    await Promise.all([
+                        mapLocationStore.resetAllMapLocations(oldForestMap),
+                        bagStore.resetBag(),
+                        chestStore.resetChest(),
+                    ]);
+
+                    chestStore.isShown = false;
                     localStorage.clear();
-                    router.push("/login");
-                });
-            });
+                    localStorage.setItem('uStatus', 'false');
+                    this.user.setLoggedIn(false);
+                    router.push('/login');
+                } catch (error) {
+                    console.error("Error during logout:", error);
+                }
+            } else {
+                console.error('Old Forest map not found');
+            }
         },
         async clearErrorMsg() {
-            this.error = "";
-        }
-    }
+            this.error = '';
+        },
+    },
 });
